move math textures to separate folder
