+++ include/slg/textures/math/add.h
