+++ include/slg/textures/math/clamp.h
