+++ include/slg/textures/math/scale.h
