+++ include/slg/textures/math/subtract.h
