+++ src/slg/scene/parsetextures.cpp
@@ -29,20 +29,16 @@
 #include "slg/scene/scene.h"
 #include "slg/utils/filenameresolver.h"
 
-#include "slg/textures/abs.h"
-#include "slg/textures/add.h"
 #include "slg/textures/band.h"
 #include "slg/textures/bilerp.h"
 #include "slg/textures/blackbody.h"
 #include "slg/textures/blender_texture.h"
 #include "slg/textures/brick.h"
 #include "slg/textures/checkerboard.h"
-#include "slg/textures/clamp.h"
 #include "slg/textures/colordepth.h"
 #include "slg/textures/constfloat.h"
 #include "slg/textures/constfloat3.h"
 #include "slg/textures/cloud.h"
-#include "slg/textures/divide.h"
 #include "slg/textures/dots.h"
 #include "slg/textures/densitygrid.h"
 #include "slg/textures/fbm.h"
@@ -60,12 +56,16 @@
 #include "slg/textures/irregulardata.h"
 #include "slg/textures/lampspectrum.h"
 #include "slg/textures/marble.h"
-#include "slg/textures/mix.h"
+#include "slg/textures/math/abs.h"
+#include "slg/textures/math/add.h"
+#include "slg/textures/math/clamp.h"
+#include "slg/textures/math/divide.h"
+#include "slg/textures/math/mix.h"
+#include "slg/textures/math/remap.h"
+#include "slg/textures/math/scale.h"
+#include "slg/textures/math/subtract.h"
 #include "slg/textures/normalmap.h"
 #include "slg/textures/object_id.h"
-#include "slg/textures/remap.h"
-#include "slg/textures/scale.h"
-#include "slg/textures/subtract.h"
 #include "slg/textures/vectormath/dotproduct.h"
 #include "slg/textures/windy.h"
 #include "slg/textures/wrinkled.h"
